+++ p/[id].tsx
@@ -1,0 +1,35 @@
+import React from "react"
+import { GetServerSideProps } from "next"
+import ReactMarkdown from "react-markdown"
+import Layout from "../components/Layout"
+import prisma from '../lib/prisma';
+
+export const getServerSideProps: GetServerSideProps = async ({ params }) => {
+  const id = Number(params?.id);
+
+  if (isNaN(id)) {
+    return {
+      notFound: true,
+    };
+  }
+
+  const listing = await prisma.listing.findUnique({
+    where: { id },
+    include: {
+      user: {
+        select: { fullName: true },
+      },
+      roommates: true,
+    },
+  });
+
+  if (!listing) {
+    return {
+      notFound: true,
+    };
+  }
+
+  return {
+    props: { listing },
+  };
+};
