--- pages/p/[id].tsx
@@ -1,76 +1,0 @@
-import React from "react"
-import { GetServerSideProps } from "next"
-import ReactMarkdown from "react-markdown"
-import Layout from "../../components/Layout"
-import { PostProps } from "../../components/Post"
-import prisma from '../../lib/prisma';
-
-export const getServerSideProps: GetServerSideProps = async ({ params }) => {
-  const id = Number(params?.id);
-
-  if (isNaN(id)) {
-    return {
-      notFound: true,
-    };
-  }
-
-  const listing = await prisma.listing.findUnique({
-    where: { id },
-    include: {
-      user: {
-        select: { fullName: true },
-      },
-      roommates: true,
-    },
-  });
-
-  if (!listing) {
-    return {
-      notFound: true,
-    };
-  }
-
-  return {
-    props: { listing },
-  };
-};
-
-const Post: React.FC<PostProps> = (props) => {
-  let title = props.title
-  if (!props.published) {
-    title = `${title} (Draft)`
-  }
-
-  return (
-    <Layout>
-      <div>
-        <h2>{title}</h2>
-        <p>By {props?.author?.name || "Unknown author"}</p>
-        <ReactMarkdown children={props.content} />
-      </div>
-      <style jsx>{`
-        .page {
-          background: white;
-          padding: 2rem;
-        }
-
-        .actions {
-          margin-top: 2rem;
-        }
-
-        button {
-          background: #ececec;
-          border: 0;
-          border-radius: 0.125rem;
-          padding: 1rem 2rem;
-        }
-
-        button + button {
-          margin-left: 1rem;
-        }
-      `}</style>
-    </Layout>
-  )
-}
-
-export default Post
